dig doesnt like "derp derp" for some reason
